+++ spec/eve_online/exceptions/timeout_spec.rb
@@ -2,6 +2,6 @@
 
 require 'spec_helper'
 
-describe EveOnline::Exceptions::TimeoutException do
+describe EveOnline::Exceptions::Timeout do
   specify { expect(subject).to be_a(EveOnline::Exceptions::Base) }
 end